signaling/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import Room, Participant

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username']

class RoomSerializer(serializers.ModelSerializer):
    class Meta:
        model = Room
        fields = ['room_id', 'host', 'created_at']

class ParticipantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Participant
        fields = ['user', 'room', 'joined_at']